feat: added code sand box exmaples
